pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kanary-prompt"
version = "0.1.0"
description = "Kanary prompt formatter registered for NeMo/Canary2 pipelines."
readme = {file = "README", content-type = "text/markdown"}
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
authors = [{name = "Kanary Project"}]
dependencies = [
    "torch",
    "lhotse",
    "nemo_toolkit",
]

[project.urls]
Homepage = "https://github.com/<your-org>/kanary_prompt"
Repository = "https://github.com/<your-org>/kanary_prompt"

[tool.setuptools.packages.find]
include = ["kanary_prompt"]
